Working version with complete feature set
The only known bug is that the camera does not start out viewing the sheet music. Pressing 1 or 2 to view the sheet music before having moved the camera, breaks the camera.
Workaround: First click in the screen to move the camera's view closer to the sheet music, then hit 1 to orient the camera. After that, 1 and 2 work as intended to view the separate sheets.